t2.26/sched_param
Bring back __sched_param to bits/sched.h

Upstream removed it because it doesn't use it, but we do: our pthread.h
includes thread-attr.h to bring the pthread_attr_t type to applications,
and thus needs the sched_param structure.

Note: that uses __sched_param and not sched_param because thread-attr.h
is also also brought in from sys/types.h, which should not expose
sched_param if not requested to by applicationg defining _POSIX_SOURCE
and alike.

Signed-off-by: Samuel Thibault <samuel.thibault@ens-lyon.org>